Fix number of pits bug + some clean up
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@2299 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: b015108f79d349fb12a8473cb1b0caa70845b242
Former-commit-id: 283d01a60d801df7dc016f7dd66c835c12e3f23a

+++ src/drivers/simplix/src/unitstrategy.cpp
@@ -132,7 +132,7 @@ bool TSimpleStrategy::IsPitFree()
 bool TSimpleStrategy::NeedPitStop()
 {
 #ifdef SPEED_DREAMS
-  double FuelConsum;                             // Fuel consumption per m
+  float FuelConsum;                              // Fuel consumption per m
   if (oFuelPerM == 0.0)                          // If still undefined
     FuelConsum = oExpectedFuelPerM;              //   use estimated value
   else                                           // If known
@@ -352,13 +352,13 @@ double TSimpleStrategy::SetFuelAtRaceStart
 
   oStartFuel =            
 	GfParmGetNum(*CarSettings,TDriver::SECT_PRIV,         // Tankinhalt beim Start
-	PRV_START_FUEL,(char*) NULL,oStartFuel);          
+	PRV_START_FUEL,(char*) NULL,(float) oStartFuel);          
   GfOut("#oStartFuel (private) = %.1f\n",oStartFuel);
 
   if ((!TDriver::Qualification)                  // Fürs Rennen 
 	  && (oStartFuel > 0))
   {
-    oLastFuel = oStartFuel;                      // volltanken
+    oLastFuel = (float) oStartFuel;              // volltanken
     GfParmSetNum(*CarSettings,SECT_CAR,PRM_FUEL, // Gewünschte Tankfüllung
       (char*) NULL, oLastFuel);                  //   an TORCS melden
     return oLastFuel;
